pyproject.toml
[tool.poetry]
name = "mosque-clock-ai"
version = "0.1.0"
description = ""
authors = ["shrhawk <shrhawk88@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.115.9"
uvicorn = "0.34.0"
openai = "1.65.1"
python-multipart = "0.0.20"
python-dotenv = "1.0.1"
pydantic = "2.10.6"
slowapi = "0.1.9"
pre-commit = "4.1.0"

[tool.poetry.dev-dependencies]
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
